ログインとサインアップのテストはaccountsのほうで実行させるようにした
